+++ test/test_nls.cpp
@@ -27,7 +27,6 @@
 
 #include <Eigen/Sparse>
 
-#include "smooth/lie_group.hpp"
 #include "smooth/optim.hpp"
 #include "smooth/so3.hpp"
